requirements.txt
streamlit
pandas
python-dotenv
groq
reportlab
beautifulsoup4
requests